Fix: auto_paging return latest fetched set of rows
If the fetched rows was the latest page, because the paging_state flag
was nil, the iteration used to stop and thus, not run the loop for the
latest page of rows.

This fix allows the loop to run one last time to expose the very last
page before actually stopping.